+++ BedeGaming.Simple.Slot.Machine/BedeGaming.SimpleSlotMachine.Application/Interfaces/Validators/IDepositValidator.cs
@@ -2,7 +2,7 @@
 
 namespace BedeGaming.SimpleSlotMachine.Application.Interfaces.Validators
 {
-    public interface IDepositValidator : IValidator<double>
+    public interface IDepositValidator : IValidator<decimal>
     {
     }
 }
